+++ Packages/Tracking/CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 - Fixed several resource leaks in `Connection`, including unmanaged memory, connection handles and event handlers.
+- Fixed erroneous "LeapProvider not assigned" log messages during teardown.
 
 ### Changed
 - Deprecated `Dispose()` on `Connection` and made it a thin shim to `Stop()` to enable better cleanup of resources.